backend/requirements.txt
connexion
python_dateutil >= 2.7.3
setuptools >= 21.0.0
flask_jwt_extended
flask_bcrypt
flask_cors
flask_mail
flask_sqlalchemy
cloudstorage
bidict
requests
celery
psycopg2-binary
sh
raven[flask]
validate_email
sqlalchemy-continuum
